Make the retrying client more generic
...so that it can be used for non-authorization purposes
