--- api/authorization/retrying_k8s_client.go
@@ -1,104 +1,0 @@
-package authorization
-
-import (
-	"context"
-	"time"
-
-	"code.cloudfoundry.org/korifi/controllers/webhooks"
-	"github.com/go-logr/logr"
-	k8serrors "k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/util/wait"
-	"k8s.io/client-go/util/retry"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-)
-
-func NewDefaultBackoff() wait.Backoff {
-	return wait.Backoff{
-		Duration: 5 * time.Millisecond,
-		Factor:   2,
-		Steps:    10,
-	}
-}
-
-type AuthRetryingClient struct {
-	client.WithWatch
-	backoff wait.Backoff
-}
-
-func NewAuthRetryingClient(c client.WithWatch, backoff wait.Backoff) client.WithWatch {
-	return AuthRetryingClient{
-		WithWatch: c,
-		backoff:   backoff,
-	}
-}
-
-func (a AuthRetryingClient) Get(ctx context.Context, key client.ObjectKey, obj client.Object, opts ...client.GetOption) error {
-	return a.retryOnForbidden(ctx, func() error {
-		return a.WithWatch.Get(ctx, key, obj, opts...)
-	}, "get")
-}
-
-func (a AuthRetryingClient) List(ctx context.Context, list client.ObjectList, opts ...client.ListOption) error {
-	return a.retryOnForbidden(ctx, func() error {
-		return a.WithWatch.List(ctx, list, opts...)
-	}, "list")
-}
-
-func (a AuthRetryingClient) Create(ctx context.Context, obj client.Object, opts ...client.CreateOption) error {
-	return a.retryOnForbidden(ctx, func() error {
-		return a.WithWatch.Create(ctx, obj, opts...)
-	}, "create")
-}
-
-func (a AuthRetryingClient) Delete(ctx context.Context, obj client.Object, opts ...client.DeleteOption) error {
-	return a.retryOnForbidden(ctx, func() error {
-		return a.WithWatch.Delete(ctx, obj, opts...)
-	}, "delete")
-}
-
-func (a AuthRetryingClient) Update(ctx context.Context, obj client.Object, opts ...client.UpdateOption) error {
-	return a.retryOnForbidden(ctx, func() error {
-		return a.WithWatch.Update(ctx, obj, opts...)
-	}, "update")
-}
-
-func (a AuthRetryingClient) Patch(ctx context.Context, obj client.Object, patch client.Patch, opts ...client.PatchOption) error {
-	return a.retryOnForbidden(ctx, func() error {
-		return a.WithWatch.Patch(ctx, obj, patch, opts...)
-	}, "patch")
-}
-
-func (a AuthRetryingClient) DeleteAllOf(ctx context.Context, obj client.Object, opts ...client.DeleteAllOfOption) error {
-	return a.retryOnForbidden(ctx, func() error {
-		return a.WithWatch.DeleteAllOf(ctx, obj, opts...)
-	}, "deleteAllOf")
-}
-
-func (a AuthRetryingClient) retryOnForbidden(ctx context.Context, fn func() error, op string) error {
-	count := 0
-	return retry.OnError(a.backoff, isForbidden, func() error {
-		logger := logr.FromContextOrDiscard(ctx).WithName("retrying-client")
-		err := fn()
-		if err != nil {
-			count++
-			logger.Info("k8s client returned forbidden", "op", op, "count", count, "error", err)
-		}
-		return err
-	})
-}
-
-// isForbidden returns true for forbidden errors that are NOT korifi webhook
-// validation errors, false otherwise upon webhook validation errors it makes
-// no sense to retry the operation as the webhook is expected to consistently
-// return the same validation error
-func isForbidden(err error) bool {
-	if !k8serrors.IsForbidden(err) {
-		return false
-	}
-
-	if _, isValidationErr := webhooks.WebhookErrorToValidationError(err); isValidationErr {
-		return false
-	}
-
-	return true
-}
